除算の場合、返り値が浮動小数点になる場合があるため、floatに変更

+++ app/Libraries/Calculation/Operator/OperatorInterface.php
@@ -6,5 +6,5 @@ namespace App\Libraries\Calculation\Operator;
 
 interface OperatorInterface
 {
-	public function calc(int $val1, int $val2): int;
+	public function calc(int $val1, int $val2): float;
 }
